Fixes for coverage testing.
Define all the environment variables at once, instead of using a matrix of
them, this wasn't the intention.

Run the tests on the "raw" sources and not on the ones in the distribution as
they don't have the same line numbers because of Pod::Weaver.

Also add coverage badge to the README as it seems to work.